+++ R/uas_thumbnails_make.R
@@ -21,9 +21,9 @@
 #' This is to prevent clashes when thumnbail files from different flights are 'gathered' into a single folder attached to
 #' a Table of Contents folder (see \code{\link{uas_toc}}).
 #'
-#' If \code{use_magick = TRUE}, it will use resizing functions from the magick package. This is slower than the equivalent
+#' If \code{use_magick = TRUE}, it will use resizing functions from the \link[magick]{magick} package. This is slower than the equivalent
 #' functions from the \code{imager} package (the default), but may be necessary if you are processing TIFs and don't have
-#' the ImageMagick app installed on your computer (which \code{imager} requires to read TIFs).
+#' \href{https://imagemagick.org}{ImageMagick} installed on your computer (which \code{imager} requires to read TIFs).
 #'
 #' @return A named list (one element for each directory processed) of thumbnail files created in the output directory
 #'
@@ -108,8 +108,9 @@ uas_thumbnails_make <- function(x, img_dir = NULL, output_dir = NULL, tb_width =
         magick_force_use <- FALSE
         if (!use_magick) {
           if ((TRUE %in% grepl(".TIF$", all_img_fn, ignore.case = TRUE)) && !imager:::has.magick()) {
-              warning("Going to use the magick package to resize TIF files. To use the (faster) imager package with TIFs, please install the ImageMagick app.")
-            magick_force_use <- TRUE
+              #warning("Going to use the magick package to resize TIF files. To use the (faster) imager package with TIFs, please install the ImageMagick app.")
+              if (!quiet) message(yellow(" - going to use the magick package to resize TIF files"))
+              magick_force_use <- TRUE
           }
         }
 
@@ -258,8 +259,7 @@ uas_thumbnails_make <- function(x, img_dir = NULL, output_dir = NULL, tb_width =
                         } else {
                             load.image(all_img_fn[j]) %>%
                                 resize(size_x = tb_width, size_y = height_new, interpolation = 3) %>%
-                                save.image(file = tb_fn[j])
-                                ## Doing 100 RGB thumbnails: 361.75 sec elapsed
+                                imager::save.image(file = tb_fn[j])
                         }
 
                         if (stats) num_tb_created <- num_tb_created + 1